+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## vNext
+
+Features:
+
+- add omitNilProperties function on AxiosClient's paramsSerializer
+
 ## v0.3.0 (July 3, 2023)
 
 Features: